// client/src/lib/database.types.ts
export interface Database {
    public: {
        Tables: {
// ...
            items: {
                Row: {
                    id: string
                    sku: string
                    name: string
                    description: string | null
                    manufacturer: string
                    model: string | null
                    serial_number: string | null
                    minimum_stock: number | null
                    unit_cost: number
                    status: 'in_stock' | 'low_stock' | 'out_of_stock' | 'discontinued'
                    created_at: string | null
                    updated_at: string | null
                    created_by: string | null
                    updated_by: string | null
                    category_id: number
                    image_url: string | null
                }
                Insert: {
                    id?: string
                    sku: string
                    name: string
                    description?: string | null
                    manufacturer: string
                    model?: string | null
                    serial_number?: string | null
                    minimum_stock?: number | null
                    unit_cost: number
                    status?: 'in_stock' | 'low_stock' | 'out_of_stock' | 'discontinued'
                    created_at?: string | null
                    updated_at?: string | null
                    created_by?: string | null
                    updated_by?: string | null
                    image_url?: string | null
                    category_id?: number
                }
                Update: {
                    id: string
                    sku?: string
                    name?: string
                    description?: string | null
                    manufacturer?: string
                    model?: string | null
                    serial_number?: string | null
                    minimum_stock?: number | null
                    unit_cost?: number
                    status?: 'in_stock' | 'low_stock' | 'out_of_stock' | 'discontinued'
                    created_at?: string | null
                    updated_at?: string | null
                    created_by?: string | null
                    updated_by?: string | null
                    category_id?: number
                    image_url?: string | null
                }
            }
            item_locations: {
                Row: {
                    id: string
                    item_id: string
                    location_id: string
                    quantity: number
                    purchased_date: string | null
                    warranty_expiration: string | null
                    is_paid: boolean | null
                    status: 'in_stock' | 'ordered'
                    created_at: string | null
                    updated_at: string | null
                    created_by: string | null
                    updated_by: string | null
                }
                Insert: {
                    id?: string
                    item_id: string
                    location_id: string
                    quantity: number
                    purchased_date?: string | null
                    warranty_expiration?: string | null
                    is_paid?: boolean | null
                    status?: 'in_stock' | 'ordered'
                    created_at?: string | null
                    updated_at?: string | null
                    created_by?: string | null
                    updated_by?: string | null
                }
                Update: {
                    id?: string
                    item_id?: string
                    location_id?: string
                    quantity?: number
                    purchased_date?: string | null
                    warranty_expiration?: string | null
                    is_paid?: boolean | null
                    status?: 'in_stock' | 'ordered'
                    created_at?: string | null
                    updated_at?: string | null
                    created_by?: string | null
                    updated_by?: string | null
                }
            }
            locations: {
                Row: {
                    id: string
                    building: string | null
                    room: string | null
                    unit: string
                    created_at: string | null
                    updated_at: string | null
                    created_by: string | null
                    updated_by: string | null
                }
                Insert: {
                    id?: string
                    building?: string | null
                    room?: string | null
                    unit: string
                    created_at?: string | null
                    updated_at?: string | null
                    created_by?: string | null
                    updated_by?: string | null
                }
                Update: {
                    id?: string
                    building?: string | null
                    room?: string | null
                    unit?: string
                    created_at?: string | null
                    updated_at?: string | null
                    created_by?: string | null
                    updated_by?: string | null
                }
            },
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
    }
}